+++ CHANGELOG.md
@@ -1,22 +1,33 @@
 # Change Log
 
-## [0.2.0]- 2021-04-17
+## [0.2.2]- 2022-04-27
+
+   - IMPROVED object support in network dataframes
+   - IMPROVED None and 'empty' handling for project metadata
+   - FIXED data types not saved for empty networks
+   - FIXED element defaults not added during element creation
+
+## [0.2.1]- 2022-04-22
+
+   - FIXED superuser flag not respected when activating project
+
+## [0.2.0]- 2022-04-17
 
    - ADDED Python 3.7 support
    - ADDED getting/setting project settings and metadata
    - ADDED full database access for superusers
    - ADDED element handling to API and client
 
-## [0.1.3]- 2021-03-31
+## [0.1.3]- 2022-03-31
 
    - FIXED wrong API routes
 
-## [0.1.2]- 2021-03-29
+## [0.1.2]- 2022-03-29
 
    - FIXED PyPi links
    - ADDED shields and description
 
-## [0.1.0]- 2021-03-29
+## [0.1.0]- 2022-03-29
 
    - Initial release